bug #4405: Optional KEEP_SNAPSHOTS parameter for VMM drivers. It preserves system snapshots across power on/off cycles and life migrations if the hypervisor supports that.

+++ src/lcm/LifeCycleActions.cc
@@ -275,7 +275,10 @@ void  LifeCycleManager::migrate_action(int vid)
 
         vm->set_resched(false);
 
-        vm->delete_snapshots();
+        if ( !vmm->is_keep_snapshots(vm->get_vmm_mad()) )
+        {
+            vm->delete_snapshots();
+        }
 
         vm->reset_info();
 
@@ -871,7 +874,10 @@ void LifeCycleManager::clean_up_vm(VirtualMachine * vm, bool dispose,
 
     vm->set_resched(false);
 
-    vm->delete_snapshots();
+    if ( !vmm->is_keep_snapshots(vm->get_vmm_mad()) )
+    {
+        vm->delete_snapshots();
+    }
 
     vm->reset_info();
